增加博客分类功能 1/2

+++ src/git/initgit.js
@@ -5,10 +5,14 @@ const config = require('../config')
 
 module.exports =
     async function init() {
+        //验证文件夹初始化
         if (!fs.existsSync('./data/.git')) {
             await gitinitSync()
         }
-        if (!fs.existsSync('./data/pathdata.sqlite')) {
+        //拉取测试
+        await gitdeal.pullSync()
+        //验证关键文件初始化
+        if (!fs.existsSync('./data/pathdata.sqlite') || !fs.existsSync('./data/markdown') || !fs.existsSync('./data/templement')) {
             await initDictionarySync()
         }
     }
@@ -19,9 +23,7 @@ function gitinitSync() {
         try {
             sg.init(false, () => {
                 sg.addRemote(config.remotename, config.gitpath, () => {
-                    sg.pull(config.remotename, 'master', () => {
-                        res()
-                    })
+                    res()
                 })
             })
         } catch (error) {
@@ -36,12 +38,19 @@ function gitinitSync() {
 function initDictionarySync() {
     return new Promise((res, rej) => {
         console.log('init start')
-        fs.writeFileSync('./data/pathdata.sqlite', '')
-        fs.mkdirSync('./data/markdown')
-        fs.writeFileSync('./data/markdown/.gitkeep', '# Dir for markdown savepath\n\n用于存放md文件的文件夹')
-        fs.mkdirSync('./data/templement')
-        fs.writeFileSync('./data/templement/.gitkeep', '# Dir for templement savepath\n\n用于存放结构体的文件夹')
-        fs.writeFileSync('./data/ReadMe.md', require('../autodata/readmefile'))
+        if (!fs.existsSync('./data/pathdata.sqlite'))
+            fs.writeFileSync('./data/pathdata.sqlite', '')
+        if (!fs.existsSync('./data/markdown')) {
+            fs.mkdirSync('./data/markdown')
+            fs.writeFileSync('./data/markdown/.gitkeep', '# Dir for markdown savepath\n\n用于存放md文件的文件夹')
+        }
+        if (!fs.existsSync('./data/templement')) {
+            fs.mkdirSync('./data/templement')
+            fs.writeFileSync('./data/templement/.gitkeep', '# Dir for templement savepath\n\n用于存放结构体的文件夹')
+        }
+        if (!fs.existsSync('./data/ReadMe.md')) {
+            fs.writeFileSync('./data/ReadMe.md', require('../autodata/readmefile'))
+        }
         sg.add('.', () => {
             sg.commit(`init at ${new Date().toUTCString()}`, () => {
                 sg.push('blobdata', 'master', () => {
@@ -53,7 +62,7 @@ function initDictionarySync() {
 }
 
 if (config.autoupdate && !config.usewebhoock) {
-    setInterval(()=>{
+    setInterval(() => {
         autoupdate()
     }, config.updattime * 60000)
 }
